Keep the composer wiring assertions in the allowlisted test file
main replaced the numeric structural-test budget with a named migration
allowlist that forbids new source-reading test files, so splitting the
wiring assertions into their own file failed the merge build twice over:
the new file was unlisted, and the file it left behind became an obsolete
entry.

Put the assertions back where the allowlist expects them. The extra
behavioral cases that motivated the split are no longer a problem, because
the allowlist tracks files rather than counting cases. The relaxed match on
the queueClipboardTextUndoably call is kept.
