+++ frontend/src/components/ChatView/__tests__/markdownClipboard.test.js
@@ -1,5 +1,6 @@
 import { test } from 'node:test'
 import assert from 'node:assert/strict'
+import { readFileSync } from 'node:fs'
 import {
   assistantClipboardText,
   insertClipboardText,
@@ -184,4 +185,21 @@ test('Markdown insertion can fall back when the browser rejects native editing',
   }, '**new**', () => { rejected += 1 }), true)
   runQueued()
   assert.equal(rejected, 1)
+})
+
+test('assistant prose and the composer share the Markdown clipboard boundary', () => {
+  const message = readFileSync(new URL('../MsgContent.jsx', import.meta.url), 'utf8')
+  const composer = readFileSync(new URL('../ChatInputBar.jsx', import.meta.url), 'utf8')
+  assert.match(message, /onCopy=\{\(event\) => copyAssistantSelection\(event, markdownForBlock\)\}/)
+  assert.match(message, /data-assistant-markdown-block=/)
+  assert.match(message, /if \(markdownByIndex\) return markdownByIndex\.get\(index\) \?\? ''/,
+    'a partial cold render must not fall back to the hidden full message source')
+  assert.match(composer, /assistantClipboardText\(\s*e\.clipboardData,\s*preferPlainText/)
+  assert.match(composer, /pasteAsPlainTextRef\.current = isPlainTextPasteShortcut\(e\)/)
+  assert.match(composer, /if \(queueClipboardTextUndoably\(/,
+    'Markdown paste should enter the browser undo stack before using the controlled fallback')
+  assert.match(composer, /pendingComposerCaretRef\.current = next/,
+    'paste and sent-message history should share the controlled caret handoff')
+  assert.doesNotMatch(composer, /setSelectionRange\(next\.caret/,
+    'paste must not introduce a second requestAnimationFrame caret path')
 })
